feat(toggle)!: toggle list on button click
Modify the JS file to add a toggle button that controls the display of
the `ul` element. This improvement streamlines the user experience and
makes it easier to hide and show the list items.

There is a known issue where the first `li` is not being displayed as
expected.

Additionally, make small changes to the index, styles, and text.json
files to better support the toggle feature.